src/PYDBtoPahoConnector/requirements.txt
paho-mqtt == 1.6.1
psycopg2 == 2.9.9